# openapi.yaml
openapi: 3.0.0
info:
  title: OkTolk API
  version: 1.0.0
paths:
  /register:
    post:
      summary: Register user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: {type: string}
                password: {type: string}
                telegram_id: {type: string}
      responses:
        '200':
          description: User created
  /login:
    post:
      summary: Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: {type: string}
                password: {type: string}
      responses:
        '200':
          description: Returns JWT token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer